requirements.txt
py-cord[voice]
openai
python-dotenv